atils/pyproject.toml
[tool.poetry]
name = "atils"
version = "0.1.0"
description = "A series of scripts to make my life easier, and automate away tedious homelab tasks."
authors = ["Aidan Hilt <aidanhilt2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mypy = "^1.1.1"
kubernetes = "^26.1.0"
argparse = "^1.4.0"
pyyaml = "^6.0"
types-pyyaml = "^6.0.12.9"
deepdiff = "^6.3.0"
shutils = "^0.1.0"
requests = "^2.29.0"
hvac = "^1.1.0"

[tool.poetry.group.dev.dependencies]
MonkeyType = "22.2.0"
black = "22.12.0"
pytest = "7.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
